adding sfdisk lvm logic to darchi.sh

+++ darchi.sh
@@ -463,32 +463,48 @@ lvm_hooks(){
 }
 
 lv_create(){
-    echo "What disk are you installing to? (nvme0n1, sda, sdb, etc)"; read disk
-    IN_DEVICE=/dev/"$disk"
-    echo "What partition is your EFI device? (nvme0n1p1, sda1, etc)"; read efi_dev
-    EFI_DEVICE=/dev/"$efi_dev"
-    echo "What partition is your Physical Device for your Volume Group? (sda2, nvme0n1p2, sdb2, etc)"; read root_dev
-    ROOT_DEVICE=/dev/"$root_dev"
     VOL_GROUP=arch_vg
     LV_ROOT="ArchRoot"
     LV_HOME="ArchHome"
     LV_SWAP="ArchSwap"
 
-    EFI_SIZE=512M
+    echo "What disk are you installing to? (nvme0n1, sda, sdb, etc)"; read disk
+    IN_DEVICE=/dev/"$disk"
+    echo "What partition is your Physical Device for your Volume Group? (sda2, nvme0n1p2, sdb2, etc)"; read root_dev
+    ROOT_DEVICE=/dev/"$root_dev"
     echo "How big is your root partition? (12G, 50G, 100G, etc)"; read rootsize
     ROOT_SIZE="$rootsize"
     echo "How big is your Swap partition?"; read swap_size
     SWAP_SIZE="$swap_size"
     #HOME_SIZE=16G
 
+    if $(efi_boot_mode); then
+        echo "What partition is your EFI device? (nvme0n1p1, sda1, etc)"; read efi_dev
+        EFI_DEVICE=/dev/"$efi_dev"
+        EFI_SIZE=512M
+        # Create the physical partitions
+        sgdisk -Z "$IN_DEVICE"
+        sgdisk -n 1::+"$EFI_SIZE" -t 1:ef00 -c 1:EFI "$IN_DEVICE"
+        sgdisk -n 2 -t 2:8e00 -c 2:VOLGROUP "$IN_DEVICE"
+
+        # Format the EFI partition
+        mkfs.fat -F32 "$EFI_DEVICE"
+    else
+        echo "What partition is your BOOT device? (nvme0n1p1, sda1, etc)"; read boot_dev
+        BOOT_DEVICE=/dev/"$boot_dev"
+        BOOT_SIZE=512M
+
+cat > /tmp/sfdisk.cmd << EOF
+$BOOT_DEVICE : start= 2048, size=+$BOOT_SIZE, type=83, bootable
+$ROOT_DEVICE : type=83
+EOF
+        # Using sfdisk because we're talking MBR disktable now...
+        sfdisk /dev/sda < /tmp/sfdisk.cmd 
+    fi
+
+
     clear
 
-    # Create the physical partitions
-    sgdisk -Z "$IN_DEVICE"
-    sgdisk -n 1::+"$EFI_SIZE" -t 1:ef00 -c 1:EFI "$IN_DEVICE"
-    sgdisk -n 2 -t 2:8e00 -c 2:VOLGROUP "$IN_DEVICE"
-    # Format the EFI partition
-    mkfs.fat -F32 "$EFI_DEVICE"
 
     # create the physical volumes
     pvcreate "$ROOT_DEVICE"
@@ -510,8 +526,8 @@ lv_create(){
     modprobe dm_mod
     # activate the vol group
     vgchange -ay
-    # format the volumes
-    mkfs.fat -F32 "$EFI_DEVICE"
+    ## format the volumes
+    #mkfs.fat -F32 "$EFI_DEVICE"
     mkfs.ext4 /dev/"$VOL_GROUP"/"$LV_ROOT"
     mkfs.ext4 /dev/"$VOL_GROUP"/"$LV_HOME"
     # mount the volumes
